Move Shopping cart interaction to seprate class
