+++ Assets/Scripts/SceneObjects/ShoppingCart/ShoppingCartInteract.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0c729dc785fc8cc4aab75c7b2b6f0e2f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
